fix(registry): guard submitter URL fallback for split attribution
Only copy authorProfileUrl to submittedByUrl when author and submitter
identities match, reuse slug-based identity helpers in attribution UI,
and apply className on provenance author fallback spans.
